simclr-physics-baseline/requirements.txt
# Core dependencies
torch
torchvision
numpy
pillow
tqdm
matplotlib
opencv-python
einops
transformers
deepspeed

